requirements.txt
# PDF Processing
pdfminer.six>=20221105
pdfplumber>=0.10.3

# Configuration
PyYAML>=6.0.1

# Core Dependencies
python-dateutil>=2.8.2
regex>=2023.10.3

# Optional Development Dependencies
black>=23.12.1
pytest>=7.4.4
flake8>=7.0.0

# Additional Dependencies
pandas>=2.0.0
openpyxl>=3.1.2  # For Excel export support
fpdf>=1.7.2

pytest==8.3.5
pandas==2.2.1
openpyxl==3.1.2
PyYAML==6.0.1
python-dateutil==2.8.2